Change SelectCommand to FindCommand, add new FindCommand class
